some changes

+++ database/migrations/2022_06_03_094218_create_stagiaires_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("nom_st");
             $table->string("prenom_st");
             $table->foreignId("groupe_id");
-            $table->integer("heure_absence_st")->default(0);
+            $table->float("heure_absence_st")->default(0);
             $table->string("numero_personnelle")->default("-----");
         });
     }
